feat: :sparkles: redesign to now account for the negation of a set being computable
The original design ignored that the opposite of a label is also computable. This is now accounted for by defining a LabelCollection with a set of atomic labels - this is necessary to know what the negation is. Also, the AtomicLabelSet was replaced by just using sets and frozensets as needed